@@ -33,7 +33,7 @@
     "migrate": "yarn generate && npx prisma migrate deploy --preview-feature",
     "start": "NODE_PATH=./dist node ./dist/index.js",
     "prepare": "",
-    "lint": "eslint --max-warnings 0 --ext .js,ts src",
+    "lint": "eslint --max-warnings 0 --ext .js,ts --cache src",
     "lint-staged-husky": "lint-staged",
     "lint-prettier": "prettier --config .prettierrc --list-different '{src,test}/**/*.{ts,js,css,scss}'",
     "prettier": "prettier --config .prettierrc --write './**/*.{ts,js,css,scss,json,md}'",
@@ -50,7 +50,7 @@
     "docker:prod": "../dockerProd.sh",
     "podman:dev": "podman-compose --file docker-compose.yml --file docker-compose.dev.yml up --build",
     "email-dev": "email dev --dir ./src/templates/emails --port 3777",
-    "check-deps": "npx npm-check",
+    "check-deps": "npx --yes npm-check",
     "sentry:sourcemaps": "sentry-cli sourcemaps inject --org desci-labs --project nodes-backend ./dist && sentry-cli sourcemaps upload --org desci-labs --project nodes-backend ./dist && echo 'Sentry sourcemaps uploaded'",
     "madge": "tsc && npx --yes madge --circular --image circularDepGraph.svg dist/index.js && open circularDepGraph.svg 2>/dev/null || xdg-open circularDepGraph.svg"
   },
@@ -146,32 +146,33 @@
     "@types/supertest": "^2.0.12",
     "@types/ws": "^8.5.12",
     "@types/xmldom": "^0.1.34",
-    "@typescript-eslint/eslint-plugin": "^6.14.0",
-    "@typescript-eslint/parser": "^6.14.0",
+    "@typescript-eslint/eslint-plugin": "^8.10.0",
+    "@typescript-eslint/parser": "^8.10.0",
     "chai": "^4.3.4",
     "copyfiles": "^2.4.1",
     "dotenv": "^10.0.0",
-    "eslint": "^7.32.0",
+    "eslint": "^8.0.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.29.0",
+    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-plugin-import-x": "^4.3.1",
     "eslint-plugin-no-array-reduce": "^1.0.62",
-    "eslint-plugin-prettier": "^5.0.1",
-    "lint-staged": "^11.1.2",
+    "eslint-plugin-prettier": "^5.2.1",
+    "lint-staged": "^15.2.10",
     "mocha": "^10.2.0",
     "nodemon": "^2.0.22",
     "nyc": "^15.1.0",
-    "prettier": "^3.1.1",
-    "pretty-quick": "^3.1.1",
+    "prettier": "^3.3.3",
+    "pretty-quick": "^4.0.0",
     "supertest": "^6.3.3",
     "ts-node": "^10.9.2",
-    "typescript": "^5.0.0"
+    "typescript": "^5.6.3"
   },
   "lint-staged": {
     "*": [
       "pretty-quick --staged"
     ],
     "{src,test}/**/*.{js,ts}": [
-      "eslint --max-warnings 0"
+      "eslint --max-warnings 0 --cache"
     ]
   },
   "prisma": {
