@@ -1,7 +1,5 @@
 import { PrismaClient } from '@prisma/client';
 
 export const prisma = new PrismaClient(
-  process.env.PRISMA_DEBUG
-    ? { log: ["info", "query", "warn", "error"] }
-    : { log: ["warn", "error"]}
+  process.env.PRISMA_DEBUG ? { log: ['info', 'query', 'warn', 'error'] } : { log: ['warn', 'error'] },
 );
