@@ -6,8 +6,9 @@ import { Request, Response } from 'express';
 import { prisma } from '../../client.js';
 import { logger as parentLogger } from '../../logger.js';
 import { ensureUniquePathsDraftTree, getLatestDriveTime } from '../../services/draftTrees.js';
-import { NodeUuid, getLatestManifestFromNode } from '../../services/manifestRepo.js';
-import repoService from '../../services/repoService.js';
+import { getLatestManifestFromNode } from '../../services/manifestRepo.js';
+import { repoService } from '../../services/repoService.js';
+import { NodeUuid } from '../../types/nodes.js';
 import { prepareDataRefsForDraftTrees } from '../../utils/dataRefTools.js';
 import { ensureUuidEndsWithDot } from '../../utils.js';
 
@@ -130,9 +131,9 @@ export const moveData = async (req: Request, res: Response<MoveResponse | ErrorR
       },
     });
 
-    const newRefs = await prepareDataRefsForDraftTrees(node.uuid, updatedManifest);
+    const newRefs = await prepareDataRefsForDraftTrees(node.uuid!, updatedManifest);
     const existingRefMap = existingDataRefs.reduce((map, ref) => {
-      map[neutralizePath(ref.path)] = ref;
+      map[neutralizePath(ref.path!)] = ref;
       return map;
     }, {});
 
