@@ -4,15 +4,10 @@ import { Response, Request } from 'express';
 
 import { prisma } from '../../client.js';
 import { logger as parentLogger } from '../../logger.js';
-import { processExternalCidDataToIpfs } from '../../services/data/externalCidProcessing.js';
+import { ExternalCid, processExternalCidDataToIpfs } from '../../services/data/externalCidProcessing.js';
 import { ensureUuidEndsWithDot } from '../../utils.js';
 
 import { ErrorResponse, UpdateResponse } from './update.js';
-
-export type ExternalCid = {
-  cid: string;
-  name: string;
-};
 
 export type ExternalCidPayload = {
   uuid: string;
