@@ -1,4 +1,5 @@
 import type { Request, Response } from 'express';
+
 import { prisma } from '../../client.js';
 import { logger as parentLogger } from '../../logger.js';
 import { ensureUuidEndsWithDot } from '../../utils.js';
@@ -54,7 +55,7 @@ export const checkNodeAccess = async (
           transactionId: true,
           commitId: true,
         },
-        orderBy: { createdAt: "desc" },
+        orderBy: { createdAt: 'desc' },
       },
     },
     where: {
@@ -67,7 +68,7 @@ export const checkNodeAccess = async (
     res.status(404).send({ ok: false, message: 'Node not found' });
     return;
   }
-  
+
   const privSharedNode = !!shareId
     ? await prisma.privateShare.findFirst({
         where: {
@@ -79,8 +80,7 @@ export const checkNodeAccess = async (
 
   const isOwner = owner?.id === node.ownerId;
   const hasAccess = privSharedNode?.nodeUUID === node.uuid || isOwner;
-  const latestPublishedVersion = node.versions
-    .find(nv => nv.transactionId !== null || nv.commitId !== null);
+  const latestPublishedVersion = node.versions.find((nv) => nv.transactionId !== null || nv.commitId !== null);
   const isPublished = !!latestPublishedVersion;
 
   res.send({
