@@ -4,11 +4,10 @@ import { NextFunction, Response } from 'express';
 
 import { prisma } from '../../client.js';
 import { logger as parentLogger } from '../../logger.js';
-import { ensureNodeAccess, RequestWithNode } from '../../middleware/authorisation.js';
-import { delFromCache } from '../../redisClient.js';
+import { RequestWithNode } from '../../middleware/authorisation.js';
 import { updateManifestDataBucket } from '../../services/data/processing.js';
-import { NodeUuid } from '../../services/manifestRepo.js';
-import repoService from '../../services/repoService.js';
+import { repoService } from '../../services/repoService.js';
+import { NodeUuid } from '../../types/nodes.js';
 import { prepareDataRefsForDagSkeleton } from '../../utils/dataRefTools.js';
 import { dagifyAndAddDbTreeToIpfs } from '../../utils/draftTreeUtils.js';
 import { ensureUuidEndsWithDot } from '../../utils.js';
