@@ -2,8 +2,8 @@ import type { Request, Response, NextFunction } from 'express';
 
 import { prisma } from '../../client.js';
 import { logger as parentLogger } from '../../logger.js';
-import { NodeUuid } from '../../services/manifestRepo.js';
 import { type ThumbnailMap, thumbnailsService } from '../../services/Thumbnails.js';
+import { NodeUuid } from '../../types/nodes.js';
 import { ensureUuidEndsWithDot } from '../../utils.js';
 
 type ThumbnailsReqBodyParams = {
