@@ -1,10 +1,10 @@
 import { Request, Response, NextFunction } from 'express';
 import jwt from 'jsonwebtoken';
 
-import { oneYear } from '../controllers/auth/magic.js';
 import { JwtPayload } from '../types/JwtPayload.js';
 import { createJwtToken } from '../utils/createJwtToken.js';
 import { CustomError } from '../utils/response/custom-error/CustomError.js';
+import { oneYear } from '../utils/time.js';
 
 export const checkJwt = (req: Request, res: Response, next: NextFunction) => {
   let token: string | undefined;
