@@ -1,18 +1,6 @@
-
-import {
-  Body,
-  Container,
-  Column,
-  Head,
-  Html,
-  Img,
-  Link,
-  Row,
-  Section,
-  Text,
-  Font,
-} from '@react-email/components';
+import { Body, Container, Column, Head, Html, Img, Link, Row, Section, Text, Font } from '@react-email/components';
 import React from 'react';
+
 import { BaseProvider } from './BaseProvider.js';
 
 export const emailAssetsBaseUrl = 'https://pub.desci.com/ipfs';
@@ -51,7 +39,7 @@ const MainLayout = ({ children, footerMsg = defaultFooterMsg }: { children: JSX.
             <Container className="backdrop-blur-2xl bg-opacity-50">
               <Section className="h-full backdrop-blur-lg w-full" align="center">
                 <Img
-                  src={ `${emailAssetsBaseUrl}/${labsLogo}` }
+                  src={`${emailAssetsBaseUrl}/${labsLogo}`}
                   width="193"
                   height="60"
                   alt="Desci Labs"
@@ -66,7 +54,7 @@ const MainLayout = ({ children, footerMsg = defaultFooterMsg }: { children: JSX.
                 <Column className="">
                   <Link href="https://desci.com" target="_blank" rel="noopener noreferrer">
                     <Img
-                      src={ `${emailAssetsBaseUrl}/${labsLogo}` }
+                      src={`${emailAssetsBaseUrl}/${labsLogo}`}
                       width="135"
                       height="42"
                       alt="Desci Labs"
