@@ -1,11 +1,16 @@
 import { ResearchObjectV1 } from '@desci-labs/desci-models';
 import { Node } from '@prisma/client';
 import axios from 'axios';
+import { errWithCause } from 'pino-std-serializers';
 
 import { PUBLIC_IPFS_PATH } from '../config/index.js';
 import { logger as parentLogger } from '../logger.js';
-import { hexToCid, isCid } from '../utils.js';
 import { getOrCache } from '../redisClient.js';
+import { hexToCid, isCid } from '../utils.js';
+
+const logger = parentLogger.child({
+  module: 'utils:manifest.ts',
+});
 
 const IPFS_RESOLVER_OVERRIDE = process.env.IPFS_RESOLVER_OVERRIDE;
 
@@ -13,7 +18,7 @@ export const cleanupManifestUrl = (url: string, gateway?: string) => {
   if (url && (PUBLIC_IPFS_PATH || gateway)) {
     const s = url.split('/');
     const res = `${gateway ? gateway : PUBLIC_IPFS_PATH}/${s[s.length - 1]}`;
-    parentLogger.info({ fn: 'cleanupManifestUrl', url, gateway }, `resolving ${url} => ${res}`);
+    logger.info({ fn: 'cleanupManifestUrl', url, gateway }, `resolving ${url} => ${res}`);
     return res;
   }
   return url;
@@ -35,36 +40,28 @@ export const transformManifestWithHistory = (data: ResearchObjectV1, researchNod
 };
 
 /** Resolve manifest given its CID, in either hex or plain-text format */
-export const resolveNodeManifest = async (
-  targetCid: string,
-  gateway?: string
-) => {
+export const resolveNodeManifest = async (targetCid: string, gateway?: string) => {
   const ipfsResolver = IPFS_RESOLVER_OVERRIDE || gateway || 'https://ipfs.desci.com/ipfs';
   let cidString = targetCid;
 
   if (!isCid(targetCid)) {
     cidString = hexToCid(targetCid);
-  };
+  }
 
   try {
-    parentLogger.info(`Calling IPFS Resolver ${ipfsResolver} for CID ${cidString}`);
+    logger.info(`Calling IPFS Resolver ${ipfsResolver} for CID ${cidString}`);
     const { data } = await axios.get(`${ipfsResolver}/${cidString}`);
     return data;
   } catch (err) {
-    // res.status(500).send({ ok: false, msg: 'ipfs uplink failed, try setting ?g= querystring to resolver' });
+    logger.error({ err: errWithCause(err) }, 'Failed to call IPFS resolver');
     return null;
   }
 };
 
-export const cachedGetDpidFromManifest = async (
-  cid: string,
-  gateway?: string
-) => {
+export const cachedGetDpidFromManifest = async (cid: string, gateway?: string) => {
   const fnGetDpidFromManifest = async () => {
-    const manifest = await resolveNodeManifest(cid, gateway) as ResearchObjectV1;
-    return manifest.dpid
-      ? parseInt(manifest.dpid.id)
-      : -1;
+    const manifest = (await resolveNodeManifest(cid, gateway)) as ResearchObjectV1;
+    return manifest.dpid ? parseInt(manifest.dpid.id) : -1;
   };
 
   const manifestDpid = await getOrCache(`manifest-dpid-${cid}`, fnGetDpidFromManifest);
@@ -73,6 +70,6 @@ export const cachedGetDpidFromManifest = async (
   } else {
     return manifestDpid;
   }
-}
+};
 
 export const zeropad = (data: string) => (data.length < 2 ? `0${data}` : data);
