@@ -1,7 +1,8 @@
 import { type Response } from 'express';
 
-import { oneDay, oneMinute, oneYear } from '../controllers/auth/magic.js';
 import { logger } from '../logger.js';
+
+import { oneDay, oneMinute, oneYear } from './time.js';
 
 /**
  * To enable a wildcard auth cookie that works across all subdomains, we need to modify the auth cookie name for each domain.
